feat: adding the ideia when the product is paused, and the option field
On the creation of a new product now have the field that you can type
and press enter the options for creating an array of options. On the all
products page now when the product is paused the image will be gray and
will show a message that the product is unavaible.